requirements.txt
pillow==10.4.0
pyqt5==5.15.11